Criei Estado Global
