+++ projeto-pokedex/src/Pages/PokedexPage/PokedexPage.js
@@ -1,23 +1,24 @@
-import React,{useState} from 'react'
+import React,{useContext} from 'react'
 import PokemonCard from '../../Componentes/PokemonCard/PokemonCard';
 import { PokedexContainer } from './pokedexPageStyle';
 import Header from '../../Componentes/Header/Header';
 import { Link } from 'react-router-dom';
+import GlobalContext from '../../contexts/GlobalContext';
 
-const PokedexPage = ({pokedex,removeFromPokedex}) => {
+const PokedexPage = () => {
   // const [pokedex, setPokedex] = useState([]);
-
+  const context = useContext(GlobalContext);
   return (
     <>
       <div>
       <Header leftButton={<Link to="/">Voltar para Lista de pokemons</Link>} />
       </div>
     <PokedexContainer>
-      {pokedex.map(pokemon => (
+      {context.pokedex.map(pokemon => (
         <PokemonCard
           key={pokemon.name}
           pokemon={pokemon}
-          removeFromPokedex={removeFromPokedex}
+          removeFromPokedex={context.removeFromPokedex}
           activePage="/pokedex"
         />
       ))}
