+++ projeto-pokedex/src/contexts/GlobalContext.js
@@ -1,0 +1,6 @@
+import {createContext} from 'react';
+
+
+const GlobalContext = createContext();
+
+export default GlobalContext;
